+++ src/command/post_to_database.rs
@@ -7,7 +7,7 @@ use dbpath::DatabasePath;
 use dbtype::PostToDatabaseResponse;
 use docid::DocumentId;
 use docpath::DocumentPath;
-use error::{Error, ErrorResponse};
+use error::{EncodeErrorKind, Error, ErrorResponse};
 use revision::Revision;
 use transport::{self, Command, Request};
 
@@ -59,7 +59,8 @@ impl<'a, T> Command for PostToDatabase<'a, T> where T: 'a + serde::Serialize
     fn make_request(self) -> Result<(Request, Self::State), Error> {
         let db_path = self.path.clone();
         let uri = self.path.into_uri(self.client_state.uri.clone());
-        let body = try!(serde_json::to_vec(self.doc_content).map_err(|e| Error::Encode { cause: e }));
+        let body = try!(serde_json::to_vec(self.doc_content)
+                            .map_err(|e| Error::Encode(EncodeErrorKind::Serde { cause: e })));
         let req = try!(Request::new(hyper::method::Method::Post, uri))
                       .accept_application_json()
                       .content_type_application_json()
@@ -77,19 +78,13 @@ impl<'a, T> Command for PostToDatabase<'a, T> where T: 'a + serde::Serialize
                 let rev: Revision = content.rev.into();
                 Ok((rev, path))
             }
-            hyper::status::StatusCode::BadRequest => {
-                Err(Error::InvalidRequest { response: try!(ErrorResponse::from_reader(resp)) })
-            }
-            hyper::status::StatusCode::Unauthorized => {
-                Err(Error::Unauthorized { response: try!(ErrorResponse::from_reader(resp)) })
-            }
-            hyper::status::StatusCode::NotFound => {
-                Err(Error::NotFound { response: Some(try!(ErrorResponse::from_reader(resp))) })
-            }
+            hyper::status::StatusCode::BadRequest => Err(Error::InvalidRequest(try!(ErrorResponse::from_reader(resp)))),
+            hyper::status::StatusCode::Unauthorized => Err(Error::Unauthorized(try!(ErrorResponse::from_reader(resp)))),
+            hyper::status::StatusCode::NotFound => Err(Error::NotFound(Some(try!(ErrorResponse::from_reader(resp))))),
             hyper::status::StatusCode::Conflict => {
                 // Need to include this error variant in the command's
                 // documentation if we ever add support for an explicit id.
-                Err(Error::DocumentConflict { response: try!(ErrorResponse::from_reader(resp)) })
+                Err(Error::DocumentConflict(try!(ErrorResponse::from_reader(resp))))
             }
             _ => Err(Error::UnexpectedHttpStatus { got: resp.status }),
         }